refactor: move SupersetAppInitializer to specific initialization package
